Fix typo which caused error in mapping
Old code only set up a normal mode mapping, when it was supposed to set
up all 3.  This means there was no visual mode mapping, and the 'ounmap'
command below gave an error on startup.

Obviously, this is the first time I actually tested this plugin.  The
good news is that the new colorscheme options worked exactly as I'd
hoped.